+++ src/index.html
@@ -17,6 +17,7 @@
           <i class="fa-solid fa-right-to-bracket add-btn" id="add-btn"></i>        
         </li>
       </ul>
+      <span class="clear-all-list" >Clear all Completed</span>
     </section>
 </body>
 </html>
