+++ src/index.js
@@ -1,6 +1,9 @@
 /* eslint-disable max-classes-per-file */
 
 import './style.css';
+import Store from '../modules/Store';
+import Ui from '../modules/Ui';
+import List from '../modules/List';
 import '@fortawesome/fontawesome-free/js/fontawesome';
 import '@fortawesome/fontawesome-free/js/solid';
 import '@fortawesome/fontawesome-free/js/regular';
@@ -9,75 +12,6 @@ import '@fortawesome/fontawesome-free/js/brands';
 // variables
 
 const toDoList = document.querySelector('.to-do-list');
-
-// classes
-
-class List {
-  constructor(description, index) {
-    this.description = description;
-    this.completed = false;
-    this.index = index;
-  }
-}
-
-class Store {
-  static getLists = () => {
-    let Lists;
-    if (localStorage.getItem('Lists') === null) {
-      Lists = [];
-    } else {
-      Lists = JSON.parse(localStorage.getItem('Lists'));
-    }
-    return Lists;
-  }
-
-  static setLists = (list) => {
-    const Lists = this.getLists();
-    Lists.push(list);
-    localStorage.setItem('Lists', JSON.stringify(Lists));
-  }
-
-  static deleteLists = (deleted) => {
-    const Lists = this.getLists();
-    const filteredLists = Lists.filter(
-      (listItem) => listItem.index !== deleted,
-    );
-
-    // update index of list
-
-    filteredLists.forEach((filteredList, index) => {
-      filteredList.index = index + 1;
-    });
-
-    localStorage.setItem('Lists', JSON.stringify(filteredLists));
-  }
-}
-
-class Ui {
-  static addLists = (list) => {
-    const toDoList = document.querySelector('.to-do-list');
-    const newItem = document.createElement('li');
-
-    newItem.setAttribute('class', 'add-to-list');
-    newItem.setAttribute('id', list.index);
-    newItem.setAttribute('data-index', list.index);
-    newItem.innerHTML = `
-      <div data-index="${list.index}" class="checkbox">
-        <input  class="strikethrough" type="checkbox">
-        <label data-index="${list.index}">${list.description}</label>
-      </div>
-      <i data-index="${list.index}" class="fa-solid fa-trash-can remove-btn"></i>
-      `;
-    toDoList.appendChild(newItem);
-  }
-
-  static DisplayList = () => {
-    const Lists = Store.getLists();
-    Lists.forEach((list) => {
-      this.addLists(list);
-    });
-  }
-}
 
 document.addEventListener('DOMContentLoaded', Ui.DisplayList());
 
@@ -95,11 +29,6 @@ addInput.addEventListener('keypress', (e) => {
   addInput.value = '';
   return true;
 });
-
-// const addBtn = document.getElementById('#add-btn');
-// addBtn.addEventListener('click', () => {
-//   alert(0);
-// })
 
 const checkBoxDiv = document.querySelectorAll('.checkbox');
 checkBoxDiv.forEach((el, i) => {
@@ -131,4 +60,9 @@ toDoList.addEventListener('click', (e) => {
   }
   addInput.value = '';
   return true;
+});
+
+const deleteAll = document.querySelector('.clear-all-list');
+deleteAll.addEventListener('click', () => {
+  console.log(0);
 });
